Criacao filtros, exclusao e exportacao

+++ Front/Front Pizzaria Att/Pizzaria/src/app/shared/produto/produto.service.ts
@@ -1,11 +1,12 @@
 import { Produtos } from './produtos.model';
-import { HttpClient, HttpHeaders   } from '@angular/common/http';
+import { HttpClient, HttpHeaders, HttpParams } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { MatSnackBar } from '@angular/material/snack-bar'
 import { Observable, of } from 'rxjs';
 import { catchError, map, tap } from 'rxjs/operators';
 import { environment } from './../../../enviroments/enviroment'
-
+import Swal from 'sweetalert2/src/sweetalert2.js'
+import { MateriaPrimaFiltros } from './produtosFiltro.model';
 @Injectable({
   providedIn: 'root'
 })
@@ -33,6 +34,14 @@ export class ProdutoService {
     })
   }
 
+  construirMessagemSemRegistros(msg: string){
+    this.exibirmsg.open(msg, 'Fechar', {
+      duration: 3000,
+      horizontalPosition: "right",
+      verticalPosition: "top"
+    })
+  }
+
   criar(produto: Produtos) : Observable<Produtos> {
     const url = `${this.baseUrl}/Incluir`;
     return this.http.post<Produtos>(url, produto, this.httpOptions)
@@ -46,6 +55,24 @@ export class ProdutoService {
     const url = `${this.baseUrl}/Listar`;
   
     return this.http.get<Produtos[]>(url).pipe(
+      catchError((error) => {
+        console.error('Erro ao listar produtos', error);
+        return of([]);
+      })
+    );
+  }
+
+  listarFiltrado(filtro: MateriaPrimaFiltros): Observable<Produtos[]> {
+    const url = `${this.baseUrl}/ListarFiltrado`;
+    const filtroParametro = new HttpParams({fromObject: {
+      campo: filtro.campo,
+      criterio_IGUAL: filtro.criterio_IGUAL,
+      criterio_LIKE: filtro.criterio_LIKE,
+      valor_LIKE: filtro.valor_LIKE,
+      valor_IGUAL: filtro.valor_IGUAL
+    }});
+
+    return this.http.get<Produtos[]>(url, {params: filtroParametro}).pipe(
       catchError((error) => {
         console.error('Erro ao listar produtos', error);
         return of([]);
@@ -74,15 +101,60 @@ export class ProdutoService {
     );
   }
 
-  excluir(id: string) : Observable<Produtos> {
+  excluir(id: string) {
     const url = `${this.baseUrl}/Excluir/${id}`;
-    return this.http.delete<Produtos>(url).pipe(
-      catchError((error) => {
-        console.error('Erro ao excluir produto', error);
-        return of(null);
-      })
-    )
+    const swalWithBootstrapButtons = Swal.mixin({
+      buttonsStyling: true,
+      customClass: {
+        confirmButton: 'btn btn-success',
+        cancelButton: 'btn btn-danger ml-2'
+      }
+    });
+  
+    swalWithBootstrapButtons.fire({
+      title: 'Deseja Excluir?',
+      text: "Você não poderá reverter isso!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonText: '<span class="btn-label"><i class="fa fa-check"></i></span>Sim, excluir!',
+      cancelButtonText: '<span class="btn-label"><i class="fa fa-times"></i></span>Não, cancelar!',
+      reverseButtons: true
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.http.delete<Produtos>(url).pipe(
+          catchError((error) => {
+            console.error('Erro ao excluir produto', error);
+            swalWithBootstrapButtons.fire(
+              'Erro!',
+              'Ocorreu um erro ao excluir a matéria prima.',
+              'error'
+            );
+            return of(null);
+          })
+        ).subscribe(() => {
+          swalWithBootstrapButtons.fire(
+            'Excluído!',
+            'A matéria prima foi excluída com sucesso.',
+            'success',
+
+            
+          ).then(() => {
+            location.reload();
+          })
+        });
+      } else {
+        swalWithBootstrapButtons.fire(
+          'Cancelado',
+          'A exclusão da matéria prima foi cancelada.',
+          'info'
+        );
+      }
+    });
   }
+  
+  
+  
+  
   
 
   private handleError<Produtos>(operation = 'operation', result?: Produtos) {
